refactor(client): move client function to public package
