+++ huaweicloud/services/acceptance/cse/resource_huaweicloud_cse_microservice_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
 	"github.com/chnsz/golangsdk/openstack/cse/dedicated/v4/services"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/common"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/cse"
@@ -20,7 +21,7 @@ func getMicroserviceFunc(conf *config.Config, state *terraform.ResourceState) (i
 		return nil, err
 	}
 
-	client := cse.NewCustomClient(state.Primary.Attributes["connect_address"], "v4", "default")
+	client := common.NewCustomClient(state.Primary.Attributes["connect_address"], "v4", "default")
 	return services.Get(client, state.Primary.ID, token)
 }
